services.models_utils helpers breakdown for each model
