+++ django_app/services/transport_app/view_utils/view_helpers.py
@@ -1,15 +1,29 @@
-from services.transport_app.models_utils.models_helpers import (
-    get_cars_db_data, get_trains_db_data, is_route_exists,
+# from services.transport_app.models_utils.models_helpers import (
+#     get_cars_db_data, get_trains_db_data, is_route_exists,
+#     is_route_parsed_1_hour_ago
+# )
+from services.transport_app.models_utils.route_helpers import (
+    is_route_exists,
     is_route_parsed_1_hour_ago
 )
 from services.transport_app.api_utils.api_request_helpers import (
     get_cars_api_data, get_trains_api_data
 )
-from services.transport_app.models_utils.models_helpers import (
-    save_api_response_in_route_and_train_models,
+# from services.transport_app.models_utils.models_helpers import (
+#     save_api_response_in_route_and_train_models,
+#     save_api_response_in_route_and_car_models,
+#     update_api_response_in_route_and_car_models,
+#     update_api_response_in_route_and_train_models
+# )
+from services.transport_app.models_utils.car_helpers import (
     save_api_response_in_route_and_car_models,
     update_api_response_in_route_and_car_models,
-    update_api_response_in_route_and_train_models
+    get_cars_db_data,
+)
+from services.transport_app.models_utils.train_helpers import (
+    save_api_response_in_route_and_train_models,
+    update_api_response_in_route_and_train_models,
+    get_trains_db_data,
 )
